Revert "XdsClient: refactor transport code to make it injectable (#30183)"
This reverts commit bd9bc5fd3e62698b9d1673af673ac0393c4749ff.
