+++ build_autogenerated.yaml
@@ -702,7 +702,6 @@ libs:
   - src/core/ext/xds/xds_channel_args.h
   - src/core/ext/xds/xds_channel_stack_modifier.h
   - src/core/ext/xds/xds_client.h
-  - src/core/ext/xds/xds_client_grpc.h
   - src/core/ext/xds/xds_client_stats.h
   - src/core/ext/xds/xds_cluster.h
   - src/core/ext/xds/xds_cluster_specifier_plugin.h
@@ -717,8 +716,6 @@ libs:
   - src/core/ext/xds/xds_resource_type_impl.h
   - src/core/ext/xds/xds_route_config.h
   - src/core/ext/xds/xds_routing.h
-  - src/core/ext/xds/xds_transport.h
-  - src/core/ext/xds/xds_transport_grpc.h
   - src/core/lib/address_utils/parse_address.h
   - src/core/lib/address_utils/sockaddr_utils.h
   - src/core/lib/avl/avl.h
@@ -1388,7 +1385,6 @@ libs:
   - src/core/ext/xds/xds_certificate_provider.cc
   - src/core/ext/xds/xds_channel_stack_modifier.cc
   - src/core/ext/xds/xds_client.cc
-  - src/core/ext/xds/xds_client_grpc.cc
   - src/core/ext/xds/xds_client_stats.cc
   - src/core/ext/xds/xds_cluster.cc
   - src/core/ext/xds/xds_cluster_specifier_plugin.cc
@@ -1403,7 +1399,6 @@ libs:
   - src/core/ext/xds/xds_route_config.cc
   - src/core/ext/xds/xds_routing.cc
   - src/core/ext/xds/xds_server_config_fetcher.cc
-  - src/core/ext/xds/xds_transport_grpc.cc
   - src/core/lib/address_utils/parse_address.cc
   - src/core/lib/address_utils/sockaddr_utils.cc
   - src/core/lib/backoff/backoff.cc
