+++ grpc.gemspec
@@ -873,8 +873,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/xds/xds_channel_stack_modifier.h )
   s.files += %w( src/core/ext/xds/xds_client.cc )
   s.files += %w( src/core/ext/xds/xds_client.h )
-  s.files += %w( src/core/ext/xds/xds_client_grpc.cc )
-  s.files += %w( src/core/ext/xds/xds_client_grpc.h )
   s.files += %w( src/core/ext/xds/xds_client_stats.cc )
   s.files += %w( src/core/ext/xds/xds_client_stats.h )
   s.files += %w( src/core/ext/xds/xds_cluster.cc )
@@ -903,9 +901,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/xds/xds_routing.cc )
   s.files += %w( src/core/ext/xds/xds_routing.h )
   s.files += %w( src/core/ext/xds/xds_server_config_fetcher.cc )
-  s.files += %w( src/core/ext/xds/xds_transport.h )
-  s.files += %w( src/core/ext/xds/xds_transport_grpc.cc )
-  s.files += %w( src/core/ext/xds/xds_transport_grpc.h )
   s.files += %w( src/core/lib/address_utils/parse_address.cc )
   s.files += %w( src/core/lib/address_utils/parse_address.h )
   s.files += %w( src/core/lib/address_utils/sockaddr_utils.cc )
