+++ grpc.gyp
@@ -759,7 +759,6 @@
         'src/core/ext/xds/xds_certificate_provider.cc',
         'src/core/ext/xds/xds_channel_stack_modifier.cc',
         'src/core/ext/xds/xds_client.cc',
-        'src/core/ext/xds/xds_client_grpc.cc',
         'src/core/ext/xds/xds_client_stats.cc',
         'src/core/ext/xds/xds_cluster.cc',
         'src/core/ext/xds/xds_cluster_specifier_plugin.cc',
@@ -774,7 +773,6 @@
         'src/core/ext/xds/xds_route_config.cc',
         'src/core/ext/xds/xds_routing.cc',
         'src/core/ext/xds/xds_server_config_fetcher.cc',
-        'src/core/ext/xds/xds_transport_grpc.cc',
         'src/core/lib/address_utils/parse_address.cc',
         'src/core/lib/address_utils/sockaddr_utils.cc',
         'src/core/lib/backoff/backoff.cc',
