+++ src/core/ext/filters/client_channel/lb_policy/xds/xds_cluster_resolver.cc
@@ -52,7 +52,6 @@
 #include "src/core/ext/filters/client_channel/subchannel_interface.h"
 #include "src/core/ext/xds/xds_bootstrap.h"
 #include "src/core/ext/xds/xds_client.h"
-#include "src/core/ext/xds/xds_client_grpc.h"
 #include "src/core/ext/xds/xds_client_stats.h"
 #include "src/core/ext/xds/xds_endpoint.h"
 #include "src/core/ext/xds/xds_resource_type_impl.h"
@@ -1018,7 +1017,7 @@ class XdsClusterResolverLbFactory : public LoadBalancingPolicyFactory {
   OrphanablePtr<LoadBalancingPolicy> CreateLoadBalancingPolicy(
       LoadBalancingPolicy::Args args) const override {
     RefCountedPtr<XdsClient> xds_client =
-        GrpcXdsClient::GetFromChannelArgs(*args.args);
+        XdsClient::GetFromChannelArgs(*args.args);
     if (xds_client == nullptr) {
       gpr_log(GPR_ERROR,
               "XdsClient not present in channel args -- cannot instantiate "
