+++ src/core/ext/xds/xds_api.cc
@@ -258,18 +258,18 @@ void MaybeLogDiscoveryRequest(
   }
 }
 
-std::string SerializeDiscoveryRequest(
+grpc_slice SerializeDiscoveryRequest(
     const XdsEncodingContext& context,
     envoy_service_discovery_v3_DiscoveryRequest* request) {
   size_t output_length;
   char* output = envoy_service_discovery_v3_DiscoveryRequest_serialize(
       request, context.arena, &output_length);
-  return std::string(output, output_length);
+  return grpc_slice_from_copied_buffer(output, output_length);
 }
 
 }  // namespace
 
-std::string XdsApi::CreateAdsRequest(
+grpc_slice XdsApi::CreateAdsRequest(
     const XdsBootstrap::XdsServer& server, absl::string_view type_url,
     absl::string_view version, absl::string_view nonce,
     const std::vector<std::string>& resource_names, grpc_error_handle error,
@@ -356,7 +356,7 @@ void MaybeLogDiscoveryResponse(
 }  // namespace
 
 absl::Status XdsApi::ParseAdsResponse(const XdsBootstrap::XdsServer& server,
-                                      absl::string_view encoded_response,
+                                      const grpc_slice& encoded_response,
                                       AdsResponseParserInterface* parser) {
   upb::Arena arena;
   const XdsEncodingContext context = {client_,
@@ -369,7 +369,8 @@ absl::Status XdsApi::ParseAdsResponse(const XdsBootstrap::XdsServer& server,
   // Decode the response.
   const envoy_service_discovery_v3_DiscoveryResponse* response =
       envoy_service_discovery_v3_DiscoveryResponse_parse(
-          encoded_response.data(), encoded_response.size(), arena.ptr());
+          reinterpret_cast<const char*>(GRPC_SLICE_START_PTR(encoded_response)),
+          GRPC_SLICE_LENGTH(encoded_response), arena.ptr());
   // If decoding fails, report a fatal error and return.
   if (response == nullptr) {
     return absl::InvalidArgumentError("Can't decode DiscoveryResponse.");
@@ -437,18 +438,18 @@ void MaybeLogLrsRequest(
   }
 }
 
-std::string SerializeLrsRequest(
+grpc_slice SerializeLrsRequest(
     const XdsEncodingContext& context,
     const envoy_service_load_stats_v3_LoadStatsRequest* request) {
   size_t output_length;
   char* output = envoy_service_load_stats_v3_LoadStatsRequest_serialize(
       request, context.arena, &output_length);
-  return std::string(output, output_length);
+  return grpc_slice_from_copied_buffer(output, output_length);
 }
 
 }  // namespace
 
-std::string XdsApi::CreateLrsInitialRequest(
+grpc_slice XdsApi::CreateLrsInitialRequest(
     const XdsBootstrap::XdsServer& server) {
   upb::Arena arena;
   const XdsEncodingContext context = {client_,
@@ -525,7 +526,7 @@ void LocalityStatsPopulate(
 
 }  // namespace
 
-std::string XdsApi::CreateLrsRequest(
+grpc_slice XdsApi::CreateLrsRequest(
     ClusterLoadReportMap cluster_load_report_map) {
   upb::Arena arena;
   // The xDS server info is not actually needed here, so we seed it with an
@@ -596,7 +597,7 @@ std::string XdsApi::CreateLrsRequest(
   return SerializeLrsRequest(context, request);
 }
 
-grpc_error_handle XdsApi::ParseLrsResponse(absl::string_view encoded_response,
+grpc_error_handle XdsApi::ParseLrsResponse(const grpc_slice& encoded_response,
                                            bool* send_all_clusters,
                                            std::set<std::string>* cluster_names,
                                            Duration* load_reporting_interval) {
@@ -604,7 +605,8 @@ grpc_error_handle XdsApi::ParseLrsResponse(absl::string_view encoded_response,
   // Decode the response.
   const envoy_service_load_stats_v3_LoadStatsResponse* decoded_response =
       envoy_service_load_stats_v3_LoadStatsResponse_parse(
-          encoded_response.data(), encoded_response.size(), arena.ptr());
+          reinterpret_cast<const char*>(GRPC_SLICE_START_PTR(encoded_response)),
+          GRPC_SLICE_LENGTH(encoded_response), arena.ptr());
   // Parse the response.
   if (decoded_response == nullptr) {
     return GRPC_ERROR_CREATE_FROM_STATIC_STRING("Can't decode response.");
