+++ src/core/ext/xds/xds_api.h
@@ -32,6 +32,8 @@
 #include "envoy/admin/v3/config_dump.upb.h"
 #include "upb/def.hpp"
 
+#include <grpc/slice.h>
+
 #include "src/core/ext/xds/certificate_provider_store.h"
 #include "src/core/ext/xds/upb_utils.h"
 #include "src/core/ext/xds/xds_bootstrap.h"
@@ -148,29 +150,29 @@ class XdsApi {
 
   // Creates an ADS request.
   // Takes ownership of \a error.
-  std::string CreateAdsRequest(const XdsBootstrap::XdsServer& server,
-                               absl::string_view type_url,
-                               absl::string_view version,
-                               absl::string_view nonce,
-                               const std::vector<std::string>& resource_names,
-                               grpc_error_handle error, bool populate_node);
+  grpc_slice CreateAdsRequest(const XdsBootstrap::XdsServer& server,
+                              absl::string_view type_url,
+                              absl::string_view version,
+                              absl::string_view nonce,
+                              const std::vector<std::string>& resource_names,
+                              grpc_error_handle error, bool populate_node);
 
   // Returns non-OK when failing to deserialize response message.
   // Otherwise, all events are reported to the parser.
   absl::Status ParseAdsResponse(const XdsBootstrap::XdsServer& server,
-                                absl::string_view encoded_response,
+                                const grpc_slice& encoded_response,
                                 AdsResponseParserInterface* parser);
 
   // Creates an initial LRS request.
-  std::string CreateLrsInitialRequest(const XdsBootstrap::XdsServer& server);
+  grpc_slice CreateLrsInitialRequest(const XdsBootstrap::XdsServer& server);
 
   // Creates an LRS request sending a client-side load report.
-  std::string CreateLrsRequest(ClusterLoadReportMap cluster_load_report_map);
+  grpc_slice CreateLrsRequest(ClusterLoadReportMap cluster_load_report_map);
 
   // Parses the LRS response and returns \a
   // load_reporting_interval for client-side load reporting. If there is any
   // error, the output config is invalid.
-  grpc_error_handle ParseLrsResponse(absl::string_view encoded_response,
+  grpc_error_handle ParseLrsResponse(const grpc_slice& encoded_response,
                                      bool* send_all_clusters,
                                      std::set<std::string>* cluster_names,
                                      Duration* load_reporting_interval);
